refactor(dao/): 필요없는 인터페이스 삭제
더블 테스트를 하기 위해 생성했던 인터페이스 삭제

BREAKING CHNAGE: 각 DAO에 대응되는 인터페이스 삭제
as-is:
@Repository
public class GameDaoImpl implements GameDao {
    ...
}

@Repository
public class PieceDaoImpl implements PieceDao {
    ...
}

to-be:
@Repository
public class GameDao {
    ...
}

@Repository
public class PieceDao {
    ...
}
